Update for populateInputCandidates
The MATLABDatasetConverter will now populate input candidates when
MatlabNumericArray outputs are requested. It will use the ObjectIndex's
currently available Datasets as the potential candidates

This allows a WidgetModel typed on MatlabNumericArray to know that the
available Datasets should be used as choices displayed to a user when
filling in parameters for a mult-param script, for example.