Ensure ImageJ can be closed and relaunched
Simplified the static variables to just hold onto the created ImageJ
instance. We then use a helper class to listen for ContextDisposing
events, as use that information to decide if we should restart ImageJ or
not on subsequent requests from MATLAB.

With this change it's now possible to close ImageJ inside MATLAB and
restart it, and using the "ImageJ" script when ImageJ is already running
does nothing except print a status message.

This behavior is as desired for a single instance. Eventually it would
be nice to remove the static-isms and allow multiple ImageJs/contexts.